+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 # Changelog 
 
+## 3.0.2
+
+- fix!(SBTree): .insertDocument to return array on single inclusion
+- fix: improve fieldName fetching from query
+- fix: finding recursive and $in
+- tests: added some query and find tests
+- doc: primitives documentation
+- doc: improved navigation